Add an on-parse callback parameter to vega directive.
This change adds an optional on-parse parameter to the Vega directive which
takes a callback function. The onParse() function is invoked within the Vega
callback for vg.parse.spec(), providing notification that parsing has completed
and allowing the parent Angular code to access Vega's view API for the embedded
chart.